# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for Employee360, an open-source, self-hostable, multi-tenant employee platform. Includes unversioned operational endpoints (health checks) and the versioned client-facing API under /api/v1.",
        "title": "Employee360 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/dashboard/admin": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Counts of users, invitations by status, departments and positions for the caller's tenant, plus the 5 most recent invitations.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Tenant admin dashboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_delivery_http_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_types_dashboard.AdminDashboardResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_delivery_http_response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_delivery_http_response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/super-admin": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Organisation overview of the caller's tenant: tenant info, user counts (total/active/inactive/pending-invited and by role), invitations by status, departments, positions and the 5 most recent invitations. super_admin only; no cross-tenant aggregates.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Super admin dashboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_delivery_http_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_types_dashboard.SuperAdminDashboardResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_delivery_http_response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_delivery_http_response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_skryfon_employee360_backend_internal_delivery_http_response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_delivery_http_response.ErrorInfo"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_delivery_http_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_skryfon_employee360_backend_internal_delivery_http_response.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_skryfon_employee360_backend_internal_delivery_http_response.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_skryfon_employee360_backend_internal_domain_entity.InvitationStatus": {
            "type": "string",
            "enum": [
                "pending",
                "accepted",
                "revoked",
                "expired"
            ],
            "x-enum-varnames": [
                "InvitationStatusPending",
                "InvitationStatusAccepted",
                "InvitationStatusRevoked",
                "InvitationStatusExpired"
            ]
        },
...
        "github_com_skryfon_employee360_backend_internal_types_dashboard.AdminDashboardResponse": {
            "type": "object",
            "properties": {
                "departments": {
                    "type": "integer"
                },
                "invitations": {
                    "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_types_dashboard.InvitationCountsResponse"
                },
                "positions": {
                    "type": "integer"
                },
                "recent_invitations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_types_invitation.InvitationListItemResponse"
                    }
                },
                "users": {
                    "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_types_dashboard.UserCountsResponse"
                }
            }
        },
        "github_com_skryfon_employee360_backend_internal_types_dashboard.InvitationCountsResponse": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "integer"
                },
                "expired": {
                    "type": "integer"
                },
                "pending": {
                    "type": "integer"
                },
                "revoked": {
                    "type": "integer"
                }
            }
        },
        "github_com_skryfon_employee360_backend_internal_types_dashboard.RoleUserCountResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "integer"
                },
                "inactive": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_skryfon_employee360_backend_internal_types_dashboard.SuperAdminDashboardResponse": {
            "type": "object",
            "properties": {
                "departments": {
                    "type": "integer"
                },
                "invitations": {
                    "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_types_dashboard.InvitationCountsResponse"
                },
                "positions": {
                    "type": "integer"
                },
                "recent_invitations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_types_invitation.InvitationListItemResponse"
                    }
                },
                "tenant": {
                    "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_types_dashboard.TenantInfoResponse"
                },
                "users": {
                    "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_types_dashboard.UserCountsResponse"
                },
                "users_by_role": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_types_dashboard.RoleUserCountResponse"
                    }
                }
            }
        },
        "github_com_skryfon_employee360_backend_internal_types_dashboard.TenantInfoResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_skryfon_employee360_backend_internal_types_dashboard.UserCountsResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "integer"
                },
                "inactive": {
                    "type": "integer"
                },
                "pending_invited": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_skryfon_employee360_backend_internal_types_invitation.InvitationListItemResponse": {
            "type": "object",
            "properties": {
                "accepted_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "department_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invited_by": {
                    "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_types_invitation.InvitedByResponse"
                },
                "invited_on": {
                    "type": "string"
                },
                "position_id": {
                    "type": "string"
                },
                "revoked_at": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "role_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_skryfon_employee360_backend_internal_domain_entity.InvitationStatus"
                }
            }
        },
...
        "github_com_skryfon_employee360_backend_internal_types_invitation.InvitedByResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}